:bricks: establish NavbarData type in component folder
