+++ src/components/navigation/navbar/_data/navbarTypes.ts
@@ -1,0 +1,9 @@
+import { Link } from "../../../../types/link";
+
+export type NavbarData = {
+    links: NavigationLink[];
+};
+
+export type NavigationLink = Link & {
+    drop?: { href: string; label: string }[];
+};
