+++ src/types/link.ts
@@ -1,0 +1,5 @@
+export type Link = {
+    href: string;
+    label: string;
+    external?: boolean;
+};
